+++ src/r3f/components/CameraControls.jsx
@@ -2,13 +2,13 @@ import { forwardRef, useMemo, useEffect, useContext } from 'react';
 import { useThree, useFrame } from '@react-three/fiber';
 import { EnvironmentControls as EnvironmentControlsImpl, GlobeControls as GlobeControlsImpl } from '3d-tiles-renderer/three';
 import { useShallowOptions } from '../utilities/useOptions.js';
-import { TilesRendererContext } from './TilesRenderer.jsx';
+import { EllipsoidContext } from './TilesRenderer.jsx';
 import { useApplyRefs } from '../utilities/useApplyRefs.js';
 
 // Add a base component implementation for both EnvironmentControls and GlobeControls
 const ControlsBaseComponent = forwardRef( function ControlsBaseComponent( props, ref ) {
 
-	const { controlsConstructor, domElement, scene, camera, ellipsoid, ellipsoidFrame, tilesRenderer, ...rest } = props;
+	const { controlsConstructor, domElement, scene, camera, ellipsoid, ellipsoidFrame, ...rest } = props;
 
 	const [ defaultCamera ] = useThree( state => [ state.camera ] );
 	const [ gl ] = useThree( state => [ state.gl ] );
@@ -17,13 +17,12 @@ const ControlsBaseComponent = forwardRef( function ControlsBaseComponent( props,
 	const [ get ] = useThree( state => [ state.get ] );
 	const [ set ] = useThree( state => [ state.set ] );
 
-	const contextTilesRenderer = useContext( TilesRendererContext );
-	const appliedTilesRenderer = tilesRenderer || contextTilesRenderer;
+	const ellipsoidContext = useContext( EllipsoidContext );
 	const appliedCamera = camera || defaultCamera || null;
 	const appliedScene = scene || defaultScene || null;
 	const appliedDomElement = domElement || gl.domElement || null;
-	const appliedEllipsoid = ellipsoid || appliedTilesRenderer?.ellipsoid || null;
-	const appliedEllipsoidFrame = ellipsoidFrame || appliedTilesRenderer?.group || null;
+	const appliedEllipsoid = ellipsoid || ellipsoidContext?.ellipsoid || null;
+	const appliedEllipsoidFrame = ellipsoidFrame || ellipsoidContext?.frame || null;
 
 	// create a controls instance
 	const controls = useMemo( () => {